+++ remote/src/main/kotlin/com/kg/gettransfer/remote/model/ConfigsModel.kt
@@ -14,7 +14,7 @@ data class ConfigsModel(
     @SerializedName(ConfigsEntity.AVAILABLE_LOCALES) @Expose val availableLocales: List<LocaleModel>,
     @SerializedName(ConfigsEntity.PREFERRED_LOCALE) @Expose val preferredLocale: String,
     @SerializedName(ConfigsEntity.CARD_GATEWAYS) @Expose val cardGateways: CardGatewaysModel,
-    @SerializedName(ConfigsEntity.PAYMENT_COMMISSION) @Expose val paymentCommission: Double,
+    @SerializedName(ConfigsEntity.PAYMENT_COMMISSION) @Expose val paymentCommission: Float,
     @SerializedName(ConfigsEntity.SUPPORTED_CURRENCIES) @Expose val supportedCurrencies: List<CurrencyModel>,
     @SerializedName(ConfigsEntity.SUPPORTED_DISTANCE_UNITS) @Expose val supportedDistanceUnits: List<String>,
     @SerializedName(ConfigsEntity.OFFICE_PHONE) @Expose val officePhone: String,
